requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pdfplumber==0.11.0
langchain==0.2.7
langchain-community==0.2.7
sentence-transformers==3.0.1
faiss-cpu==1.8.0
transformers==4.42.4
torch==2.3.1
accelerate==0.33.0
numpy<2
pypdf
langchain-huggingface
groq
python-dotenv
